+++ app/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@moondreamsdev/dreamer-ui": "^1.6.6-test.6",
+    "@moondreamsdev/dreamer-ui": "^1.7.1-test.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
