+++ lib/src/components/accordion/index.ts
@@ -1,0 +1,4 @@
+export { Accordion } from './Accordion';
+export { AccordionItem } from './AccordionItem';
+export type { AccordionProps, AccordionOption } from './Accordion';
+export type { AccordionItemProps } from './AccordionItem';
