+++ lib/src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './accordion';
 export * from './button';
 export * from './checkbox';
 export * from './input';
