+++ lib/src/symbols/ChevronDown.tsx
@@ -1,4 +1,3 @@
-// TASK: move to lib
 export default function ChevronDown({ size = 15, color = 'currentColor' }: { size?: number; color?: string }) {
   return (
     <svg width={size} height={size} className={'inline'} viewBox='0 0 15 15' xmlns='http://www.w3.org/2000/svg'>
