+++ lib/src/symbols/index.ts
@@ -1,5 +1,6 @@
 export { default as Check } from './Check';
 export { default as CheckCircled } from './CheckCircled';
+export { default as ChevronDown } from './ChevronDown';
 export { default as ExclamationTriangle } from './ExclamationTriangle';
 export { default as EyeClosed } from './EyeClosed';
 export { default as EyeOpened } from './EyeOpened';
